my service layers

+++ src/main/java/am/aca/quiz/software/service/intefaces/QuestionService.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Service
 public interface QuestionService {
     //create
-    boolean addQuestion(String question,Enum level,int correctAnswerCount,double points,Long topicCentityId) throws SQLException;
+    boolean addQuestion(String question,Enum level,int correctAnswerCount,int points,Long topicCentityId) throws SQLException;
 
     //read
     List<QuestionEntity> getAll() throws SQLException;
